Re-create amplify project
This time I've done it properly...

- /authorise to get an access token (and get redirected to
mlhfellowbook.herokuapp.com
- /fellows to list all fellows
- /fellows/username to get a specific fellow by case-insensitive username
