+++ aws-backend/amplify/backend/function/fellowendpoint/amplify.state
@@ -2,5 +2,5 @@
   "pluginId": "amplify-nodejs-function-runtime-provider",
   "functionRuntime": "nodejs",
   "useLegacyBuild": true,
-  "defaultEditorFile": "src/index.js"
+  "defaultEditorFile": "src/app.js"
 }
