+++ aws-backend/amplify/backend/function/fellowendpoint/src/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "fellowsendpoint",
+  "name": "fellowendpoint",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
