+++ aws-backend/amplify/backend/function/githubauthtoken/amplify.state
@@ -1,0 +1,6 @@
+{
+  "pluginId": "amplify-nodejs-function-runtime-provider",
+  "functionRuntime": "nodejs",
+  "useLegacyBuild": true,
+  "defaultEditorFile": "src/app.js"
+}
