+++ aws-backend/amplify/backend/function/githubauthtoken/parameters.json
@@ -1,0 +1,1 @@
+{}
