+++ aws-backend/amplify/backend/function/githubauthtoken/src/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "fellowsendpoint",
+  "name": "githubauthtoken",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "aws-sdk": "^2.580.0",
     "aws-serverless-express": "^3.3.5",
     "body-parser": "^1.17.1",
-    "express": "^4.15.2"
+    "express": "^4.15.2",
+    "node-fetch": "^2.6.0"
   },
   "devDependencies": {},
   "scripts": {
