+++ aws-backend/amplify/backend/function/githuborgauth/function-parameters.json
@@ -1,0 +1,1 @@
+{}
