+++ aws-backend/amplify/backend/function/githuborgauth/parameters.json
@@ -1,0 +1,1 @@
+{}
