+++ aws-backend/amplify/backend/function/githuborgauth/src/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "githubauth",
-  "version": "2.0.0",
+  "name": "githuborgauth",
+  "version": "1.0.0",
   "description": "Lambda function generated by Amplify",
   "main": "index.js",
   "license": "Apache-2.0",
