+++ aws-backend/amplify/backend/storage/fellowsdb/storage-params.json
@@ -1,0 +1,1 @@
+{}
